First seperate the two wrappers
