+++ src/utils/CardEditorWrapper.tsx
@@ -1,35 +1,7 @@
 import Preact, { render } from "preact";
-import { CardContextProvider } from "../utils";
 import { HomeAssistant } from "custom-card-helpers";
 import { CacheProvider } from "@emotion/react";
 import createCache from "@emotion/cache";
-
-export type GetConfigValid<T> = (
-  config: T
-) =>
-  | { valid: true; errorMessage?: undefined }
-  | { valid: false; errorMessage: string };
-
-export class PreactWrapper<T> extends HTMLElement {
-  Card: Preact.FunctionComponent = null;
-  config: T = null;
-
-  set hass(hass) {
-    if (!this.Card) {
-      throw new Error("Preact Card is not defined");
-    }
-    render(
-      <CardContextProvider rootElement={this} hass={hass} config={this.config}>
-        <this.Card />
-      </CardContextProvider>,
-      this
-    );
-  }
-
-  getCardSize() {
-    return 1;
-  }
-}
 
 export type EditorCardProps<T> = {
   config: T;
@@ -37,7 +9,7 @@ export type EditorCardProps<T> = {
   hass: HomeAssistant;
 };
 
-export class PreactEditorWrapper<T> extends HTMLElement {
+export class CardEditorWrapper<T> extends HTMLElement {
   _config: T = null;
   Card: Preact.FunctionComponent<EditorCardProps<T>> = null;
   _hass: HomeAssistant = null;
